refactoring class name(Pr suggestions implemented)
